+++ contracts/spec_interfaces/IDelegation.sol
@@ -31,4 +31,5 @@ interface IDelegations /* is IStakeChangeNotifier */ {
 	function getDelegatedStakes(address addr) external view returns (uint256);
 	function getSelfDelegatedStake(address addr) external view returns (uint256);
 	function getDelegation(address addr) external view returns (address);
+	function getTotalDelegatedStake() external view returns (uint256) ;
 }
